+++ tests/Fixer/ImplodeCallFixerTest.php
@@ -36,7 +36,7 @@ final class ImplodeCallFixerTest extends AbstractFixerTestCase
      *
      * @dataProvider provideFixCases
      */
-    public function testFix(string $expected, string $input = null): void
+    public function testFix(string $expected, ?string $input = null): void
     {
         $this->doTest($expected, $input);
     }