+++ tests/Fixer/NoUselessClassCommentFixerTest.php
@@ -42,7 +42,7 @@ final class NoUselessClassCommentFixerTest extends AbstractFixerTestCase
      *
      * @dataProvider provideFixCases
      */
-    public function testFix(string $expected, string $input = null): void
+    public function testFix(string $expected, ?string $input = null): void
     {
         $this->doTest($expected, $input);
     }