+++ tests/Fixer/NoUselessConstructorCommentFixerTest.php
@@ -38,7 +38,7 @@ final class NoUselessConstructorCommentFixerTest extends AbstractFixerTestCase
      *
      * @dataProvider provideFixCases
      */
-    public function testFix(string $expected, string $input = null): void
+    public function testFix(string $expected, ?string $input = null): void
     {
         $this->doTest($expected, $input);
     }
@@ -51,7 +51,7 @@ final class NoUselessConstructorCommentFixerTest extends AbstractFixerTestCase
                 /**
                  * Destructor
                  */
-                 public function __constructor () {}
+                 public function __constructor() {}
              }
              ',
         ];
@@ -62,7 +62,7 @@ final class NoUselessConstructorCommentFixerTest extends AbstractFixerTestCase
                 /**
                  * Reconstructor
                  */
-                 public function __constructor () {}
+                 public function __constructor() {}
              }
              ',
         ];
@@ -72,7 +72,7 @@ final class NoUselessConstructorCommentFixerTest extends AbstractFixerTestCase
             class Foo {
                 /**
                  */
-                 public function __constructor () {}
+                 public function __constructor() {}
              }
              ',
             '<?php
@@ -80,7 +80,7 @@ final class NoUselessConstructorCommentFixerTest extends AbstractFixerTestCase
                 /**
                  * Constructor
                  */
-                 public function __constructor () {}
+                 public function __constructor() {}
              }
              ',
         ];
@@ -90,7 +90,7 @@ final class NoUselessConstructorCommentFixerTest extends AbstractFixerTestCase
             class Foo {
                 /**
                  */
-                 public function __constructor () {}
+                 public function __constructor() {}
              }
              ',
             '<?php
@@ -98,7 +98,7 @@ final class NoUselessConstructorCommentFixerTest extends AbstractFixerTestCase
                 /**
                  * constructor
                  */
-                 public function __constructor () {}
+                 public function __constructor() {}
              }
              ',
         ];
@@ -108,7 +108,7 @@ final class NoUselessConstructorCommentFixerTest extends AbstractFixerTestCase
             class Foo {
                 /**
                  */
-                 public function __constructor () {}
+                 public function __constructor() {}
              }
              ',
             '<?php
@@ -116,7 +116,7 @@ final class NoUselessConstructorCommentFixerTest extends AbstractFixerTestCase
                 /**
                  * Constructor.
                  */
-                 public function __constructor () {}
+                 public function __constructor() {}
              }
              ',
         ];
@@ -126,7 +126,7 @@ final class NoUselessConstructorCommentFixerTest extends AbstractFixerTestCase
             class Foo {
                 /**
                  */
-                 public function __constructor () {}
+                 public function __constructor() {}
              }
              ',
             '<?php
@@ -134,7 +134,7 @@ final class NoUselessConstructorCommentFixerTest extends AbstractFixerTestCase
                 /**
                  * Foo Constructor
                  */
-                 public function __constructor () {}
+                 public function __constructor() {}
              }
              ',
         ];
@@ -144,7 +144,7 @@ final class NoUselessConstructorCommentFixerTest extends AbstractFixerTestCase
             class Foo {
                 /**
                  */
-                 public function __constructor () {}
+                 public function __constructor() {}
              }
              ',
             '<?php
@@ -152,7 +152,7 @@ final class NoUselessConstructorCommentFixerTest extends AbstractFixerTestCase
                 /**
                  * FooBar Constructor
                  */
-                 public function __constructor () {}
+                 public function __constructor() {}
              }
              ',
         ];
@@ -162,7 +162,7 @@ final class NoUselessConstructorCommentFixerTest extends AbstractFixerTestCase
             class Foo {
                 /**
                  */
-                 public function __constructor () {}
+                 public function __constructor() {}
              }
              ',
             '<?php
@@ -170,7 +170,7 @@ final class NoUselessConstructorCommentFixerTest extends AbstractFixerTestCase
                 /**
                  * Foo\Bar Constructor
                  */
-                 public function __constructor () {}
+                 public function __constructor() {}
              }
              ',
         ];
@@ -182,7 +182,7 @@ final class NoUselessConstructorCommentFixerTest extends AbstractFixerTestCase
                  *
                  * This class has awesome Constructor
                  */
-                 public function __constructor () {}
+                 public function __constructor() {}
              }
              ',
             '<?php
@@ -192,7 +192,7 @@ final class NoUselessConstructorCommentFixerTest extends AbstractFixerTestCase
                  *
                  * This class has awesome Constructor
                  */
-                 public function __constructor () {}
+                 public function __constructor() {}
              }
              ',
         ];
@@ -204,7 +204,7 @@ final class NoUselessConstructorCommentFixerTest extends AbstractFixerTestCase
                  * This class has awesome Constructor.
                  *
                  */
-                 public function __constructor () {}
+                 public function __constructor() {}
              }
              ',
             '<?php
@@ -214,7 +214,7 @@ final class NoUselessConstructorCommentFixerTest extends AbstractFixerTestCase
                  *
                  * Foo Constructor
                  */
-                 public function __constructor () {}
+                 public function __constructor() {}
              }
              ',
         ];
@@ -226,7 +226,7 @@ final class NoUselessConstructorCommentFixerTest extends AbstractFixerTestCase
                  * @author John Doe
                  * This class has awesome Constructor
                  */
-                 public function __constructor () {}
+                 public function __constructor() {}
              }
              ',
             '<?php
@@ -236,7 +236,7 @@ final class NoUselessConstructorCommentFixerTest extends AbstractFixerTestCase
                  * Foo Constructor.
                  * This class has awesome Constructor
                  */
-                 public function __constructor () {}
+                 public function __constructor() {}
              }
              ',
         ];
@@ -247,7 +247,7 @@ final class NoUselessConstructorCommentFixerTest extends AbstractFixerTestCase
                 /**
                  * @see example.com
                  */
-                 public function __constructor () {}
+                 public function __constructor() {}
              }
              ',
             '<?php
@@ -255,7 +255,7 @@ final class NoUselessConstructorCommentFixerTest extends AbstractFixerTestCase
                 /** Foo Constructor
                  * @see example.com
                  */
-                 public function __constructor () {}
+                 public function __constructor() {}
              }
              ',
         ];
@@ -265,14 +265,14 @@ final class NoUselessConstructorCommentFixerTest extends AbstractFixerTestCase
             class Foo {
                 //
                 // This class has awesome Constructor.
-                 public function __constructor () {}
+                 public function __constructor() {}
              }
              ',
             '<?php
             class Foo {
                 // Foo Constructor
                 // This class has awesome Constructor.
-                 public function __constructor () {}
+                 public function __constructor() {}
              }
              ',
         ];
@@ -281,7 +281,7 @@ final class NoUselessConstructorCommentFixerTest extends AbstractFixerTestCase
             '<?php
             class Foo {
                 // Foo is constructor
-                 public function __constructor () {}
+                 public function __constructor() {}
              }
              ',
         ];