+++ CHANGELOG.md
@@ -1,9 +1,17 @@
 # CHANGELOG for PHP CS Fixer: custom fixers
 
 ## [Unreleased]
+- Add NoCommentedOutCodeFixer
 - Add NoUselessCommentFixer
+- Add NullableParamStyleFixer
 - Deprecate NoUselessClassCommentFixer
 - Deprecate NoUselessConstructorCommentFixer
+- Feature: OperatorLinebreakFixer - handle ternary operator
+- Fix: NoImportFromGlobalNamespaceFixer - class without  namespace
+- Fix: NoUselessClassCommentFixer - comment detection
+- Fix: TokenRemover - remove last element of file
+- Fix: TokenRemover - remove item in line after code
+- Fix: NoImportFromGlobalNamespaceFixer - constant named the same as global imported class
 
 ## v1.11.0 - *2018-10-14*
 - Add PhpdocParamOrderFixer